requirements.txt
aliyun-python-sdk-core-v3==2.9.1
aliyun-python-sdk-kms==2.5.0
asn1crypto==0.24.0
attrs==18.2.0
Automat==0.7.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
constantly==15.1.0
crcmod==1.7
cryptography==2.3.1
cssselect==1.0.3
cx-Oracle==7.0.0
decorator==4.3.0
hyperlink==18.0.0
idna==2.7
incremental==17.5.0
jsonpath-rw==1.4.0
lxml==4.2.5
mysql-connector-python==8.0.12
oss2==2.5.0
parsel==1.5.0
ply==3.11
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
pycparser==2.18
pycryptodome==3.6.6
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pyOpenSSL==18.0.0
python-scrapyd-api==2.1.2
queuelib==1.5.0
redis==2.10.6
requests==2.19.1
Scrapy==1.5.1
scrapyd==1.2.0
selenium==3.14.0
service-identity==17.0.0
six==1.11.0
Twisted==18.7.0
urllib3==1.23
w3lib==1.19.0
xlrd==1.1.0
zope.interface==4.5.0
